backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "google-genai>=1.64.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "ruff>=0.15.1",
    "sqlalchemy>=2.0.46",
    "ty>=0.0.17",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

    [tool.coverage.run]
    source = ["src"]
    omit = ["*/tests/*", "*/__pycache__/*", "*/htmlcov/*", "*/.venv/*"]

    [tool.coverage.report]
    exclude_lines = [
        "pragma: no cover",
        "def __repr__",
        "if __name__ == .__main__.:",
        "raise AssertionError",
        "raise NotImplementedError",
    ]

    [dependency-groups]
    dev = [
        "pytest-cov>=7.0.0",
    ]